b_CUo13Nvv8fG-1775235049799/backend/requirements.txt
# Backend Requirements
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Data Parsing
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==5.1.0
feedparser==6.0.11

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
